requirements.txt
functions-framework>=3.2.0
pandas>=2.1.1
requests>=2.32.0
google-cloud-bigquery>=3.14.0
